+++ src/main/java/music/settings/PublicSettings.java
@@ -1,4 +1,4 @@
-package musicclient.model.impl;
+package music.settings;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;