+++ src/main/java/musicclient/MusicClient.java
@@ -5,9 +5,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 
 @EnableZuulProxy
 @SpringBootApplication
+@ComponentScan({"music", "musicclient"})
 public class MusicClient {
 
     public static void main(String[] args) {