+++ src/main/java/musicclient/endpoint/SettingsEndpoint.java
@@ -1,6 +1,6 @@
 package musicclient.endpoint;
 
-import musicclient.model.impl.PublicSettings;
+import music.settings.PublicSettings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;