+++ src/main/java/musicclient/endpoint/TrackEndpoint.java
@@ -1,17 +1,18 @@
 package musicclient.endpoint;
 
+import music.service.MetadataService;
 import musicclient.service.TrackService;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
+import org.jaudiotagger.tag.datatype.Artwork;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import static music.utils.EndpointUtils.responseEntity;
 
 import java.io.File;
 import java.io.IOException;
@@ -20,8 +21,14 @@ import java.io.IOException;
 @RequestMapping("/track")
 public class TrackEndpoint {
 
-    @Autowired
-    private TrackService trackService;
+    private final TrackService trackService;
+
+    private final MetadataService metadataService;
+
+    public TrackEndpoint(TrackService trackService, MetadataService metadataService) {
+        this.trackService = trackService;
+        this.metadataService = metadataService;
+    }
 
     @GetMapping("/{id}/stream")
     public ResponseEntity<Resource> stream(@PathVariable long id) throws IOException {
@@ -31,5 +38,12 @@ public class TrackEndpoint {
                 .header(HttpHeaders.CONTENT_TYPE, "audio/" + FilenameUtils.getExtension(file.getAbsolutePath()).toLowerCase())
                 .body(new InputStreamResource(FileUtils.openInputStream(file)));
 
+    }
+
+    @GetMapping("/{id}/art")
+    public ResponseEntity<Resource> getAlbumArt(@PathVariable long id, @RequestParam(defaultValue = "0") Integer index) throws IOException {
+        File file = trackService.getFile(id, true);
+        Artwork art = metadataService.getAlbumArt(file, index);
+        return responseEntity(null, art.getMimeType(), art.getBinaryData());
     }
 }